move test file to corresponding package
