+++ app/src/androidTest/java/hsk/practice/myvoca/room/vocabulary/ObjectConvertTest.kt
@@ -1,11 +1,10 @@
-package hsk.practice.myvoca.room
+package hsk.practice.myvoca.room.vocabulary
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.gson.Gson
 import hsk.practice.myvoca.data.MeaningImpl
 import hsk.practice.myvoca.data.VocabularyImpl
 import hsk.practice.myvoca.data.WordClassImpl
-import hsk.practice.myvoca.room.vocabulary.*
 import junit.framework.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
