+++ app/src/androidTest/java/hsk/practice/myvoca/room/vocabulary/SQLiteDatabaseTestHelper.kt
@@ -1,4 +1,4 @@
-package hsk.practice.myvoca.room
+package hsk.practice.myvoca.room.vocabulary
 
 class SQLiteDatabaseTestHelper {
