+++ app/src/androidTest/java/hsk/practice/myvoca/room/vocabulary/VocaDaoTest.kt
@@ -1,4 +1,4 @@
-package hsk.practice.myvoca.room
+package hsk.practice.myvoca.room.vocabulary
 
 import android.content.Context
 import androidx.room.Room
@@ -8,9 +8,7 @@ import hsk.practice.myvoca.data.MeaningImpl
 import hsk.practice.myvoca.data.VocabularyImpl
 import hsk.practice.myvoca.data.WordClassImpl
 import hsk.practice.myvoca.data.fakeData
-import hsk.practice.myvoca.room.vocabulary.VocaDao
-import hsk.practice.myvoca.room.vocabulary.toRoomVocabulary
-import hsk.practice.myvoca.room.vocabulary.toVocabularyImpl
+import hsk.practice.myvoca.room.RoomVocaDatabase
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
 import org.junit.After
